+++ Juyoung/app/src/main/java/com/example/project6/SearchPage/ChargerSearchRequest.java
@@ -1,6 +1,6 @@
-package com.example.project6.logIn;
+package com.example.project6.SearchPage;
 
-import com.example.project6.User.UserDTO;
+import android.content.ContentValues;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -12,16 +12,56 @@ import java.io.OutputStream;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.Map;
 
-public class LoginRequest {
-    public String requestLogin(String _url, UserDTO userDTO){
-        int check=400;
+public class ChargerSearchRequest {
+    public String oneOrMoreChargerSearchRequest(String _url, ContentValues params){
         HttpURLConnection urlConn= null;
+        StringBuffer sbParams =new StringBuffer();
+
+        if(params==null)
+            sbParams.append("");
+
+        try {
+            URL url = new URL(_url);
+            urlConn = (HttpURLConnection) url.openConnection();
+            urlConn.setRequestMethod("GET");
+            urlConn.setRequestProperty("Accept-Charset", "UTP - 8");
+            urlConn.setRequestProperty("ConText_Type", "application/json");
+            urlConn.setDoInput(true);
+
+            BufferedReader br = new BufferedReader(new InputStreamReader(urlConn.getInputStream(), "UTF-8"));
+            if (urlConn.getResponseCode() != HttpURLConnection.HTTP_OK)
+                return null;
+            String line;
+            String page = "";
+            while ((line = br.readLine()) != null) {
+                page += line;
+
+            }
+            return page;
+        }
+        catch (MalformedURLException e){
+            e.printStackTrace();
+        }
+        catch (IOException e){
+            e.printStackTrace();
+        }
+        finally {
+            if(urlConn!=null)
+                urlConn.disconnect();
+        }
+        return null;
+    }
+
+
+    public String searchChargerDetailRequest(String _url, String Address){
+        HttpURLConnection urlConn= null;
+        StringBuffer sbParams =new StringBuffer();
         String json="";
         JSONObject jsonObject=new JSONObject();
         try {
-            jsonObject.accumulate("email", userDTO.getEmail());
-            jsonObject.accumulate("userPass", userDTO.getPassword());
+            jsonObject.accumulate("address", Address);
         } catch (JSONException e) {
             e.printStackTrace();
         }
@@ -41,10 +81,10 @@ public class LoginRequest {
             os.flush();
             os.close();
 
-            check=urlConn.getResponseCode();
+            int check=urlConn.getResponseCode();
             BufferedReader br = new BufferedReader(new InputStreamReader(urlConn.getInputStream(), "UTF-8"));
             if (urlConn.getResponseCode() != HttpURLConnection.HTTP_OK)
-                return String.valueOf(check);
+                return null;
             String line;
             String page = "";
             while ((line = br.readLine()) != null) {
@@ -62,6 +102,9 @@ public class LoginRequest {
             if(urlConn!=null)
                 urlConn.disconnect();
         }
-        return String.valueOf(check);
+        return null;
     }
+
+
+
 }
