--- Juyoung/app/src/main/res/layout/signup.xml
@@ -1,130 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".SignUp.SignupActivity"
-    android:orientation="vertical">
-
-  <LinearLayout
-      android:layout_marginTop="10dp"
-      android:layout_marginBottom="150dp"
-      android:gravity="center"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content">
-    <TextView
-        android:layout_gravity="center"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="회원가입"
-        android:textSize="50dp" />
-  </LinearLayout>
-
-  <LinearLayout
-      android:layout_marginBottom="15dp"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content">
-    <LinearLayout
-        android:gravity="center"
-        android:orientation="horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-      <TextView
-          android:text="EMAIL"
-          android:layout_marginRight="20dp"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"/>
-      <EditText
-          android:hint="이메일을 입력하세요."
-          android:id="@+id/email_editText_inSignUp"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"/>
-    </LinearLayout>
-
-  </LinearLayout>
-
-  <LinearLayout
-      android:layout_marginBottom="15dp"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content">
-    <LinearLayout
-        android:gravity="center"
-        android:orientation="horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-      <TextView
-          android:text="PASSWORD"
-          android:layout_marginRight="15dp"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"/>
-      <EditText
-          android:hint="비밀번호를 입력하세요."
-          android:id="@+id/password_editText_inSignUp"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"/>
-    </LinearLayout>
-
-  </LinearLayout>
-
-  <LinearLayout
-      android:layout_marginBottom="15dp"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content">
-    <LinearLayout
-        android:gravity="center"
-        android:orientation="horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-      <TextView
-          android:text="NAME"
-          android:layout_marginRight="20dp"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"/>
-      <EditText
-          android:hint="이름을 입력하세요."
-          android:id="@+id/name_editText"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"/>
-    </LinearLayout>
-
-  </LinearLayout>
-
-  <LinearLayout
-      android:layout_marginBottom="50dp"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content">
-    <LinearLayout
-        android:gravity="center"
-        android:orientation="horizontal"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-      <TextView
-          android:text="PHONE"
-          android:layout_marginRight="20dp"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"/>
-      <EditText
-          android:hint="핸드폰 번호를 입력하세요."
-          android:id="@+id/phone_editText"
-          android:layout_width="wrap_content"
-          android:layout_height="wrap_content"/>
-    </LinearLayout>
-
-  </LinearLayout>
-
-  <LinearLayout
-      android:gravity="center"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content">
-    <Button
-        android:id="@+id/signUp_btn_inSignUp"
-        android:text="회원가입"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content">
-    </Button>
-  </LinearLayout>
-
-
-  </LinearLayout>
